requirements.txt
beautifulsoup4==4.11.2
pymongo==4.7.1
python-dotenv==1.0.1
requests_html==0.10.0
